drm_hwcomposer: Split FB creation and buffer import.
We currently import buffer for a given gralloc handle and also
immediately create an FB for it. We may or may not end up using
this buffer on a separate plane (layers could be squashed or need
to fallback to GL composition). Hence, split this up and create
FB for buffer only when we know that this is going to be used for
scanout. We pass in plane_type which will be taken into use
in future patches to determine the framebuffer formats to be used
for Primary on IA.

BUG=None.
TEST=No Graphics regressions observed on Skylake platform.

Change-Id: I522b055ce6c989bdc04d48f2ad9b1791b31bceab
Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>